Changed rss_scrape code to return a more structured dictionary.
